perf: 无锁双缓冲 + TCP批量读 + PSRAM池复用
并发:
- mutex→双缓冲+atomic指针交换, Core 0写back/Core 1读front, 零阻塞
- memory_order_acquire/release确保SMP内存一致性

网络:
- readBytes批量TCP读取(512B缓冲), 替代逐char read减少WiFi驱动开销
- TCP Keep-Alive(5s空闲/5s间隔/3次探测)检测死连接
- reconnect前先disconnect, 防止socket泄漏
- WiFiClient超时10秒配置

内存:
- PixelPlayer PSRAM静态池复用, 避免反复malloc/free堆碎片
- size_t显式转换, 防止uint16_t乘法溢出(大分辨率场景)

可观测:
- 堆水位监控(每5s): Internal+PSRAM free/maxAlloc
- FPS计数器(每秒): 实际渲染帧率输出

+++ esp32_firmware/src/pixel_player.h
@@ -88,7 +88,8 @@ private:
 
     // 辅助方法
     bool validateHeader(const uint8_t* data, size_t len) const;
-    size_t getFrameSize() const { return _header.width * _header.height * 2; }
+    // [Step 6] 显式size_t转换，防止uint16_t乘法溢出
+    size_t getFrameSize() const { return (size_t)_header.width * (size_t)_header.height * 2; }
     bool rleDecompress(const uint8_t* compressed, size_t compLen, uint16_t* output, size_t pixelCount);
 };
